requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
aniso8601==10.0.1
backports.entry-points-selectable==1.3.0; python_version < '3.10'
bcrypt==4.3.0; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==5.0.0
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==23.0.0; python_version >= '3.7'
honcho==2.0.0
importlib-metadata==8.5.0; python_version < '3.10'
importlib-resources==6.4.5; python_version < '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
packaging==25.0; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.2
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.40; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
typing-extensions==4.13.2; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.20.2; python_version >= '3.8'
flask-session==0.8.0; python_version >= '3.8'
